requirements.txt
# Core testing framework
pytest>=7.0.0
pytest-html>=3.1.0
pytest-xdist>=2.5.0

# Selenium and WebDriver management
selenium>=4.0.0
webdriver-manager>=3.8.0

# Test reporting and documentation
allure-pytest>=2.10.0
pytest-cov>=4.0.0

# Configuration and environment
python-dotenv>=0.19.0

# Optional: Additional testing utilities
pytest-timeout>=2.1.0
pytest-mock>=3.6.0
pytest-rerunfailures>=10.0

# Optional: For API testing (if needed later)
requests>=2.28.0

# Optional: For data handling
pandas>=1.3.0